requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
audtorch==0.6.4
alembic==1.12.0
amqp==5.1.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.14.0
atpublic==4.0
attrs==23.1.0
audeer==1.20.1
audiofile==1.3.0
audioread==3.0.1
audmath==1.4.0
audmetric==1.2.0
audobject==0.7.9
autopep8==2.0.4
billiard==4.1.0
blinker==1.6.3
celery==5.3.4
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.2.1
colorama==0.4.6
configobj==5.0.8
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
databricks-cli==0.18.0
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.8.0
docker==6.1.3
dpath==2.1.6
dulwich==0.21.6
dvc==3.26.1
dvc-data==2.18.1
dvc-http==2.30.2
dvc-objects==1.0.1
dvc-render==0.6.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
efficientnet-pytorch==0.7.1
entrypoints==0.4
filelock==3.9.0
flask==2.3.3
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
funcy==2.0
gitdb==4.0.10
GitPython==3.1.37
grandalf==0.8
greenlet==3.0.0
gto==1.4.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.2
latex==0.7.0
lazy-loader==0.3
librosa==0.10.1
llvmlite==0.41.0
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.0
mccabe==0.7.0
mdurl==0.1.2
mlflow==2.7.1
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
networkx==3.0
numba==0.58.0
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
openai-whisper==20231117
opensmile==2.5.0
orjson==3.9.9
oyaml==1.0
packaging==23.2
pandas==2.1.1
pathspec==0.11.2
Pillow==9.3.0
platformdirs==3.11.0
pooch==1.7.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
pyarrow==13.0.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.12
pydot==1.4.2
pyflakes==3.1.0
pygit2==1.13.1
Pygments==2.16.1
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2023.10.3
requests==2.28.1
rich==13.6.0
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
scmrepo==1.3.1
seaborn==0.13.0
semver==3.0.2
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.22
sqlparse==0.4.4
sqltrie==0.8.0
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tokenizers==0.14.1
tomli==2.0.1
tomlkit==0.12.1
torch==2.1.0
torchaudio==2.1.0
torchinfo==1.8.0
torchlibrosa==0.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
typer==0.9.0
tzdata==2023.3
urllib3==1.26.13
vine==5.0.0
voluptuous==0.13.1
waitress==2.1.2
wcwidth==0.2.8
websocket-client==1.6.4
werkzeug==3.0.0
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.17.0
